pyproject.toml
[tool.poetry]
name = "digitalbreak24-hack"
version = "0.1.0"
description = ""
authors = ["artemmmichurin <amichurin@sbdagroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "*"
requests = "*"
Levenshtein = "*"
psycopg = "^3.2.1"
starlette = "^0.38.4"
fastapi = "^0.114.0"
pydub = "^0.25.1"
uvicorn = "^0.30.6"
pydantic-settings = "*"
python-multipart = "^0.0.9"
ffmpeg = "*"
python-telegram-bot = "*"
transformers = "^4.44.2"
torch = "^2.4.1"
langdetect = "^1.0.9"
speechbox = "^0.2.1"
pyannote-audio = "*"
langchain_community = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
